tdf#130857 qt weld: Split QtInstanceMessageDialog::buttonForResponseCode
Split the existing logic from
QtInstanceMessageDialog::buttonForResponseCode into 2 methods:
Extract a static helper method that takes a list of buttons as
an additional parameter and move that one to the QtInstanceDialog
base class, so it can be reused there in an upcoming commit.

No longer explicitly make sure to run the
QtInstanceMessageDialog::buttonForResponseCode logic
in the main thread in that class, as that already happens
in both callers.

Change-Id: I0cbc73798caf4925496b2b26be73913dae4a89b7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182043
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>