vcl: simplify Animation::Invert()
Regardless of whether all the frames are inverted, the animation bitmap
will always be inverted. Thus, we can invert the bitmap first, and then
start inverting all the frames. This simplifies matters because we can
just exit the function with a false return if the invert fails on any
animation frame.

Secondly, this allows us to convert to a for range loop.

Change-Id: Ifa8152e9903bfc53634b67c8933b50915e7a194f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182050
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>