Avoid slight jumping of TextEdit on activation
A PixelSnap was used before using GetOutputDevice() and
LogicToPixel/PixelToLogic (what was incorrect, would need
to take care of 1/2 pixel in logic for rounding). We do not
need that anymore, in fact it leads to the text slightly
'jumping' around by up to 1 pixel (of course).

We paint text nowadays using decomposed TextPrimitives
with sub-pixel precision and similar (using a shortcut)
for text in TextEdit on the Overlay, also using sub-pixel
precision. Just remove this to avoid Text being displayed
different in TextEdit and EditView paint visualizations
and assign given value to OutputArea unchanged.

If this should cause problems somewhere (I do not think so,
we try to paint text with Sub-Pixel quality now anyways)
this can also be done in alternative ways and only for
TextEdit on Overlay.

Change-Id: If177ceec880a29a6d102ad4f0dddf891d95b69c2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182094
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>