valueset a11y: Fix reported position
ValueSetAcc::getBounds needs to report its bounding
box within its parent.

However, so far the returnded rectangle would always
have a position of (0,0), which is wrong if the ValueSet
isn't positioned at the top left of its parent.

The ValueSetAcc::getLocationOnScreen logic to get
the position relative to the parent and add the parent's
absolute position is correct in theory, but was wrong
because the position relative to the parent was incorrect.

Instead, use weld::DrawingArea::get_accessible_location_on_screen
for the ValueSet's DrawingArea to get the location on screen.

In ValueSetAcc::getBounds, subtract the screen location of
the parent to convert the absolute position to one relative
in the parent (the opposite of what ValueSetAcc::getLocationOnScreen
was doing before this commit).

Without this, an incorrect position was e.g. seen
for the following use case when using the qt6 VCL plugin:

1) in Writer, open the "Format" -> "Paragraph" dialog
2) go to the "Borders" tab
3) start Accerciser
4) in Accerciser's treeview of the LO a11y, navigate to the
   "Shadow Style" panel

-> The correct area is highlighted on screen

5) click on its "Position" label child

-> The correct area is highlighted on screen.

6) click on its list child

-> The wrong area is highlighted on screen: The highlight
rectangle's left edge is at the "Position" label again and
it doesn't reach the end of the different shadow style (ValueSet)
items.

With this commit in place, the correct area is highlighted.

Change-Id: Id31a321c6bec2a3ef9a1cbb2e5264eb6deabc57f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182112
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>