valueset a11y: Simplify ValueItemAcc::getAccessibleIndexInParent
Instead of calling ValueSetItem::GetAccessible
on various ValueSetItems (first in a loop, and the first item later)
in order to compare whether their returned accessible is self (`this`),
instead check whether the ValueSetItem is
`ValueItemAcc::mpValueSetItem`, i.e. the ValueSetItem for which
`this` is the accessible object.

Before and after this change, the reported child index is as expected
in both, Writer's "Format" -> "Paragraph" dialog's "Borders" tab (see

    Change-Id: Id31a321c6bec2a3ef9a1cbb2e5264eb6deabc57f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Feb 24 17:17:34 2025 +0100

        value set a11y: Fix reported position

for more details on how to get to the object of interest
in Accerciser.) and in Calc's "Format" -> "Cells" dialog, "Background" tab
when enabling the "Color" option when testing with the qt6 VCL plugin.
(The single color rectangles in the palette are ValueSetItems,
and when querying their index in the parent in Accerciser's IPython console,
it matches, i.e. is n for the n-th child of the ValueSet.)

Change-Id: I502c4e2271c65dc4d4c35d1f22e247b0b6ea98b9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182113
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins