tdf#164907 sw: Calculate row height adds border if 'atLeast' set #2
The idea in 25.2 commit 4f5b896d004af023f210f235aeafa9abc50068d2
was to add the border line's width to the atLeast row height.
However, the function used also included the border spacing,
which is not (really) correct.

[It appears that the BOTTOM border padding should be added,
 but that is true for both the atLeast and the 'exact' cases,
 so that change should be handled separately.]

In MS Word, the bottom border of one row
is also the top border of the next row (IsCollapsedBorders()).
The bottom border line width seems to be irrelvant
in regards to the height of the row,
so the fact that the code only adds the top border seems correct.
The last row puts the bottom border _below_ the row.

make CppunitTest_sw_layoutwriter4 \
    CPPUNIT_TEST_NAME=TestTdf164907_rowHeightAtLeast

Change-Id: I756d38109a724d06e6230f5607b80a0bdc059bc6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182139
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>