valueset a11y: Use OAccessibleComponentHelper for ValueSetAcc
Similar to how

    commit 7107db3fe773629cc511eb5922bc9c28c9c5c60a
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jan 24 18:57:58 2025 +0100

        toolkit a11y: Use OAccessibleComponentHelper for grid control

did for AccessibleGridControlBase, also derive
from OAccessibleComponentHelper for ValueSetAcc
to make use of the logic already implemented there
instead of having to implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

ValueSetAcc::getBounds implements what is needed
to implement OAccessibleComponentHelper::implGetBounds,
so rename the method (and drop the mutex guard from that
method, as callers of that method in the base class take
care of that).

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the OAccessibleComponentHelper base class.

No change in behavior intended or observed when
testing this with Accerciser and Orca and the qt6
VCL plugin for the scenario described in

    commit cfe63bf67f4a2c9943f5089781d5f26f78266913
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Feb 24 17:17:34 2025 +0100

        valueset a11y: Fix reported position

Change-Id: I59643f970a54a585c9ec1f8a5ca915014f1e2991
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182158
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins