valueset a11y: Use OAccessibleComponentHelper for ValueItemAcc
Similar to how

Change-Id: I59643f970a54a585c9ec1f8a5ca915014f1e2991
Author: Michael Weghorn <m.weghorn@posteo.de>
Date:   Tue Feb 25 11:51:40 2025 +0100

    valueset a11y: Use OAccessibleComponentHelper for ValueSetAcc

did for ValueSetAcc, also derive
from OAccessibleComponentHelper for ValueItemAcc
to make use of the logic already implemented there
instead of having to implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

ValueItemAcc::getBounds implements what is needed
to implement OAccessibleComponentHelper::implGetBounds,
so rename the method (and drop the mutex guard from that
method, as callers of that method in the base class take
care of that).

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the OAccessibleComponentHelper base class.

No change in behavior intended or observed initially
when testing this with Accerciser and Orca and the qt6
VCL plugin for the scenario described in

    commit cfe63bf67f4a2c9943f5089781d5f26f78266913
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Feb 24 17:17:34 2025 +0100

        valueset a11y: Fix reported position

However, when closing LO, this now triggers an assertion:

    soffice.bin: /home/michi/development/git/libreoffice/comphelper/source/misc/accessibleeventnotifier.cxx:142: bool (anonymous namespace)::implLookupClient(const AccessibleEventNotifier::TClientId, ClientMap::iterator &): Assertion `rClients.end() != rPos && "AccessibleEventNotifier::implLookupClient: invalid client id " "(did you register your client?)!"' failed.

This is due to a preexisting issue that ValuItemAcc
objects are not disposed that will be fixed separately
in upcoming commit

    Change-Id: Ifa7e18393edcc1889bcb390fa453c611d9345bdc
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 25 12:33:21 2025 +0100

        valueset a11y: Dispose ValueItemAcc objects

The issue didn't trigger a crash on exit before
this commit because AccessibleEventNotifier and
the static `ClientMap gaClients` it uses weren't
used before with ValueSetAcc's custom a11y event
handling.
(Considering to get rid of the static client map
might be something to further consider in the future,
but I think that makes most sense once more classes
currently using AccessibleEventNotifier directly
have been refactored to use OAccessibleComponentHelper).

The fact that Orca currently doesn't announce the
valueset items when moving focus between them is independent
of this commit. Suggested upstream fix to make it work:
https://gitlab.gnome.org/GNOME/orca/-/merge_requests/236

Change-Id: If448008b3a6dc7b22a06b6ed551b08a40b2d5de3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182173
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins