thumbnailview a11y: Don't confuse x/y pos for item location
Fix ThumbnailViewItemAcc::getLocationOnScreen by using
the y coordinate of the parent's screen position instead
of the x one when calculating the item's y position.

This makes Accerciser highlight the correct area
on screen when selecting the a11y object of one of the
recent documents in the "Recent Documents" tab in the
LO Start Center.

Change-Id: I34ad1797e042d93bebf8dca1025c914db6dda254
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182211
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>