tdf#163616 add Match Diacritics to the find bar
Trying to replicate the behaviour implemented
in the search dialog, while trying to respect
the previous logic for search bar which
ignores diacritics and kashida if CTL font
is enabled.
In my opinion, the new checkbox should be the
only parameter ruling whether the search
should match diacritics or not, but we
can still ignore kashida if CTL font is
enabled.

Change-Id: I1b88be26dd80a1ebbcf78c32ade7d5a289e0766c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182212
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>