thumbnailview a11y: Use OAccessibleComponentHelper for ThumbnailViewItemAcc
Similar to how

    commit e7dc7ebc9e14447a4ab75502a378f632a910c394
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 25 12:14:24 2025 +0100

        valueset a11y: Use OAccessibleComponentHelper for ValueItemAcc

did for ValueItemAcc, also derive
from OAccessibleComponentHelper for ThumbnailViewItemAcc
to make use of the logic already implemented there
instead of having to implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

ThumbnailViewItemAcc::getBounds implements what is needed
to implement OAccessibleComponentHelper::implGetBounds,
so rename the method (and drop the mutex guard from that
method, as callers of that method in the base class take
care of that).

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the OAccessibleComponentHelper base class.

No change in behavior intended or observed initially
when testing this with the ThumbnailView for "Recent Documents"
in the LO Start Center with Accerciser and Orca and the qt6
VCL plugin.

However, when closing LO, this now triggers an assertion:

    soffice.bin: /home/michi/development/git/libreoffice/comphelper/source/misc/accessibleeventnotifier.cxx:142: bool (anonymous namespace)::implLookupClient(const AccessibleEventNotifier::TClientId, ClientMap::iterator &): Assertion `rClients.end() != rPos && "AccessibleEventNotifier::implLookupClient: invalid client id " "(did you register your client?)!"' failed.

This is the same preexisting issue as already described
in more detail for ValueItemAcc in the above-mentioned
commit (items don't get disposed) and will be addressed
in an upcoming commit.

Change-Id: Ie4e5c300ff6bdb160e756e7173c45c0f143fac69
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182225
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>