sd slidesorter a11y: Use OAccessibleComponentHelper
Similar to how

    commit b5d197cf4b4fc5f4e6aca5e01e2cc9757293c63e
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 25 11:51:40 2025 +0100

        valueset a11y: Use OAccessibleComponentHelper for ValueSetAcc

did for ValueSetAcc, also derive
from OAccessibleComponentHelper for AccessibleSlideSorterView
to make use of the logic already implemented there
instead of having to implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

AccessibleSlideSorterView::getBounds implements what is needed
to implement OAccessibleComponentHelper::implGetBounds,
so rename the method (and drop the mutex guard from that
method, as callers of that method in the base class take
care of that).

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the OAccessibleComponentHelper base class.

Calling AccessibleSlideSorterView::Destroyed
from the dtor (which would notify listeners
that the object is disposed is no more needed
as the base class OCommonAccessibleComponent::disposing
does that, and the OCommonAccessibleComponent dtor
calls OCommonAccessibleComponent::ensureDisposed.

No change in behavior intended or observed when testing
this with Accerciser and Orca and the qt6 VCL plugin
with the "Slides" pane in Impress.

Change-Id: Ic7e6ff1450ced8fb88eca57d607dc27d6c0b9e01
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182228
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins