sd slidesorter a11y: Use OAccessibleComponentHelper for items
Similar to how

    commit b5d197cf4b4fc5f4e6aca5e01e2cc9757293c63e
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 25 11:51:40 2025 +0100

        valueset a11y: Use OAccessibleComponentHelper for ValueSetAcc

did for ValueSetAcc, also derive
from OAccessibleComponentHelper for AccessibleSlideSorterObject
to make use of the logic already implemented there
instead of having to implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

AccessibleSlideSorterObject::getBounds implements what is needed
to implement OAccessibleComponentHelper::implGetBounds,
so rename the method (and drop the mutex guard from that
method, as callers of that method in the base class take
care of that).

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the OAccessibleComponentHelper base class.

No change in behavior intended or observed when testing
this with Accerciser and Orca and the qt6 VCL plugin
with the "Slides" pane in Impress.

Change-Id: I7a7fe4c3e82632e46ac5406e366ee4763f239ec4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182248
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins