tdf#165132 DOCX import: fix unexpected rotation on entirely vertical line
Open the bugdoc, it should have a fake left paragraph border (a vertical
line), it shows up as horizontal.

This went wrong in commit 6c09c85ec384e88c89bff0817e7fe9889d7ed68e
(tdf#161779 DOCX import, drawingML: fix handling of translation for
lines, 2024-06-26) where it tried to model the ODF import, but there the
PolyPolygon and Transformation properties are set separately (in this
order) on the line shape, while oox/ code sets it via
SvxShape::setPropertyValues() and that leads to a wrong result.

Fix the problem by switching to set the two properties one by one (and
not in one go) at least for vertical line shapes, which leads to the
correct result.

Note that this is a minimal regression fix, a better (but more risky)
fix would be to improve SvxShape so setting all properties in one go
would lead to the same result. Tests which were failing while working on
this change:
- CppunitTest_oox_drawingml's testToplevelLineHorOffsetDOCX: manual
  testing shows this is still OK.
- CppunitTest_sw_ooxmlexport3's testArrowPosition and
  CppunitTest_svx_unit's testPageViewDrawLayerClip: restrict this to
  entirely vertical lines for now.

Change-Id: Ia0e113d476d785d9320a142bdb449843f61d5dfe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182207
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins