oox: export embedded font used in a presentation document
This adds support to export the fonts that are used in an
presentation document. The fonts are converted to EOT fonts.

Adds an EOT converter, which converts a TTF font to EOT font by
writing a EOT header and adding the TTF font data.

Adds BigEndianTypes so we can read a structure that is defined in
big endian and the types get converted (if needed) to system
endian when accessed. + test

Adds TTF reader which reads the true type font structure in an
easy way that is used by the EOTConverter. + test

Change-Id: I8b0bcb69e47943c2a8679edf881d7c2bab45c450
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184288
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit d0ee08cfbf12145027eee7ad46448a8734693c06)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184815
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>