sd presenter a11y: No longer read a11y names from config
Instead of reading strings for the a11y names to use
for the Impress Presenter Console objects, use the
usual approach for translatable strings.

This is simpler and I see no reason to have those
accessible names configurable.

Maybe the previous approach was somehow related
to the fact that Presenter Console used to be
an extension before

    commit ea91c7d90d74e1ca039ba669b5d3e14fa359c0fa
    Author: Stephan Bergmann <sbergman@redhat.com>
    Date:   Wed Nov 21 17:19:28 2012 +0100

        Turn presenter screen from bundled extension to plain code

Change-Id: I55b5e525913f429ec60a1c21b7bf162870d79af9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182325
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>