sd presenter a11y: Fix parent-relative paragraph position
AccessibleParagraph used to override both,
AccessibleObject::GetRelativeLocation
and AccessibleObject::GetAbsoluteParentLocation.

What was odd is that AccessibleParagraph::GetAbsoluteParentLocation
was not returning the absolute location of the actual parent,
but of the grandparent (parent's parent), because
AccessibleObject::GetRelativeLocation didn't actually
return the position relative to the direct parent, but to the
grandparent as well.

While this evens out in AccessibleObject::getLocationOnScreen
where both of these are called, it doesn't in
AccessibleObject::getLocation where only
AccessibleObject::GetRelativeLocation gets called.

Adjust the logic to let AccessibleObject::GetRelativeLocation
return the paragraph's relative location without adding to
that.

This allows to drop AccessibleParagraph::GetAbsoluteParentLocation
altogether because the base class implementation can be used as
is now.

Without this commit, the incorrect area would be highlighted
by Accerciser when using the gtk3 VCL plugin and the scenario
decribed in previous commit

    Change-Id: I74660a07e5660a0f1ac0cd17eff793e4e4a8979b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 27 17:00:44 2025 +0100

        sd presenter a11y: Avoid invalid string index

. With this commit in place, the correct area is highlighted
on screen now.

Change-Id: I1481cbb03e99b6f89c5222ed260477004a9db62c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182327
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins