sd presenter a11y: Use OAccessibleComponentHelper
Similar to how

    commit 1c3d0bc25f1a8bc506fecca7bf39d78f7f817ad3
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Feb 26 11:14:04 2025 +0100

        thumbnailview a11y: Use OAccessibleComponentHelper for ThumbnailViewItemAcc

did for ThumbnailViewItemAcc, also derive
from OAccessibleComponentHelper for AccessibleObject
to make use of the logic already implemented there
instead of having to implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

AccessibleObject::getBounds implements what is needed
to implement OAccessibleComponentHelper::implGetBounds,
so rename the method.

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the OAccessibleComponentHelper base class.

No change in behavior intended or observed initially
when testing this with the qt6 VCL plugin and the scenario
otherwise described in

    Change-Id: I74660a07e5660a0f1ac0cd17eff793e4e4a8979b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 27 17:00:44 2025 +0100

        sd presenter a11y: Avoid invalid string index

However, when closing LO, this now triggers an assertion:

    soffice.bin: /home/michi/development/git/libreoffice/comphelper/source/misc/accessibleeventnotifier.cxx:142: bool (anonymous namespace)::implLookupClient(const AccessibleEventNotifier::TClientId, ClientMap::iterator &): Assertion `rClients.end() != rPos && "AccessibleEventNotifier::implLookupClient: invalid client id " "(did you register your client?)!"' failed.

This is the same preexisting issue as already described
in more detail for ValueItemAcc in

        commit e7dc7ebc9e14447a4ab75502a378f632a910c394
        Author: Michael Weghorn <m.weghorn@posteo.de>
        Date:   Tue Feb 25 12:14:24 2025 +0100

            valueset a11y: Use OAccessibleComponentHelper for ValueItemAcc

(items don't get disposed) and will be addressed
in an upcoming commit.

Change-Id: I72ac37ad368b7d4abf14f625004dad60b20ffe81
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182328
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins