sw: layout: fix truncation of table in section
There is a table in a section, split across 2 pages; now insert a page
break that moves the section to page 2.

In SwTabFrame::MakeAll(), on page 2, first the follow is joined, which
grows the table frame but shrinks the section frame (removing the
follow's height), then the table's upper is grown to LONG_MAX but with
bTst=true so it doesn't actually change or invalidate the section frame,
then in lcl_RecalcRow() almost no growth happens because the table
frame already has accounted for the height of most of the table content.

The section frame remains too small for the table, and the content is
truncated... this problem can already be reproduced in OOo 3.3.

When GetUpper()->Grow() returns some space, invalidate the upper if it's
a section frame, then the next round will format the section.

(Simply calling Grow() with bTst=false results in an opposite problem,
where the section is too large.)

Change-Id: I3cb3a5a38fbd7996fa9fcea9a6036f4260451a95
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182331
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit 154b58cfd76a8ad670d4aa25874eed66e3398a3e)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182354
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>