tdf#134864 speedup load of pathological conditional formats in XLS
This associated file had 300k conditional formatting entries, which
reduced to about 10 when de-duplicated.

Rework the logic so instead of deduplicating rule model entries, we dedplicate at a slightly higher level, and make the comparison check more thoroughly, which greatly increases the number of duplicates we find.

Takes the load time from 38sec to 6sec on my machine.

Tweak one of the unit tests, because we now consolidate some
of the conditional formatting entries in that test file.

Change-Id: Iff2c6afb1c9959576aa00a38d0a8e5d4a48b0fd5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182356
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins