ScriptForge (dialogs, documents, widgets) Secure listeners
Listeners programmed in Basic are created with
  CreateUnoListener(prefix, listener)

The prefix argument determines the series of
routines run when an event is triggered.

So far, for SF internal listeners, the prefix
was chosen to minimize the risk of naming
collisions with routines present in user scripts.

In fact the prefix argument may be provided as
  Library.Module.Prefix
which eliminates the risk of ambiguity or
homonymy.

With the actual change forcing the full
qualification of prefixes, the risk is reduced
from epsilon to zero.

No impact on the user documentation.

Change-Id: I95ef4beec80c89a074c8bec809f416479e64d8cf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182384
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins