weld::Scrollbar: Make clear that signal is about value change
A GtkAdjustment [1] doesn't only have a current value, but
also lower + upper bounds,...

Rename weld::Scrollbar::signal_adjustment_changed
to weld::Scrollbar::signal_adjustment_value_changed
etc. to make clearer that this signal is only emitted
when the current value changes, not any of the other
things an adjustment manages (and that can be set
via the weld::Scrollbar API, for example).

See how the GtkInstanceScrollbar ctor connects
the `signalAdjustValueChanged` callback to the
GtkAdjustment:value-changed signal [2], not the
GtkAdjustment:changed signal [3], which means that
only changes to the current value trigger the callback
(which in turn calls `signal_adjustment_value_changed()`),
but changes to other GtkAdjustment properties don't.

[1] https://docs.gtk.org/gtk3/class.Adjustment.html
[2] https://docs.gtk.org/gtk3/signal.Adjustment.value-changed.html
[3] https://docs.gtk.org/gtk3/signal.Adjustment.changed.html

Change-Id: Icbeb92f3e53e0747154038aeca2f11626623fbae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182400
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins