sw: workaround crash when deleting textbox of shape
... that is positioned off-page.

Crashes in SwNode::RemoveAnchoredFly because m_pAnchoredFlys is null,
apparently the text frame was not registered at the node.

This stops crashing in 7.1 with commit
fd88bb81236ad6d1a63436b7ce20f2f93410d359 "tdf#107225 tdf#122887 sw: fix
reanchoring of textboxes" but that is a fix for 2 other commits that
aren't in this branch and has a follow-up later that fixes some problem
it apparently introduced...

For this old branch, use less risky approach of just ignoring the
problem, which avoids crashing.

Change-Id: If05fed24762400a3e3b8db23c3ce684c13b31b41