tdf#163989 vcl a11y: Rework toolbar item window a11y
If a toolbar item has a window
(as returned by ToolBox::GetItemWindow), then the
toolbar item's accessible (a VCLXAccessibleToolBoxItem)
should be reported as the parent of the accessible
object for that item window.

Reporting the item window's accessible as a child is
pretty straightforward and already implemented in
The VCLXAccessibleToolBoxItem::getAccessibleChild.

Ensuring that the item window's accessible reports
the VCLXAccessibleToolBoxItem as its parent and
the proper index is more challenging, because it
does not match the vcl::Window hierarchy. (There
is no vcl::Window for toolbar items, but only for
the toolbar itself.)

So far, in order to make that work, OToolBoxWindowItem
was the XAccessible implementation for toolbar item windows
and OToolBoxWindowItemContext was the XAccessibleContext
implementation that got returned by OToolBoxWindowItem::getAccessibleContext
(via the base class implementation OAccessibleWrapper::getAccessibleContext).

Those were subclassing OAccessibleWrapper and
OAccessibleContextWrapper in order to report
the parent (and index in parent) as needed.

One effect of this however is that the gtk3 logic
to include native GtkWidgets that are contained in a
vcl::Window in the a11y hierarchy, as implemented in

    commit 305c6fee0be4db38023d9ca5f7915e443e0bc1fc
    Author: Caolán McNamara <caolanm@redhat.com>
    Date:   Wed Mar 24 11:33:42 2021 +0000

        tdf#141197 if we have a sysobj child then include that in the atk hierarchy

does not work for that case, because it relies on
the accessible to be an XWindow in order to extract
the underlying vcl::Window. With OAccessibleWrapper
being used, that is not the case.
As a consequence, native GTK widgets inside the vcl ToolBox
(i.e. non-native toolbar) like the Paragraph Styles or Font Style
comboboxes or the combobox in Writer's search toolbar
cannot be found in Accerciser's treeview of the LO a11y hierarchy.
(There are panels, but the comboboxes inside the panels are missing.)

While the logic inside the gtk3 a11y bridge could be
extended to cover that case also (by making use of
implementation details in OAccessibleWrapper, e.g.
getting its OAccessibleWrapper::m_xInnerAccessible),
choose another solution that doesn't depend on making
use of more implementation details and ports away
from OAccessible(Context)Wrapper for the
toolbar item windows completely:

Introduce Window::SetAccessibleParent that allows
to explicitly set an accessible parent for a vcl::Window.
If this is done, Window::GetAccessibleParent, used by
VCLXAccessibleComponent::getAccessibleParent returns
this accessible instead of using the accessible for
the vcl::Window returned by vcl::Window::GetAccessibleParentWindow.

This ensures that the item window returns the
toolbar item as its parent in
VCLXAccessibleComponent::getAccessibleParent
and the base class implementation in
OCommonAccessibleComponent::getAccessibleIndexInParent
ensures that the correct child index is also returned.

Drop the now unused classes OToolBoxWindowItem and
OToolBoxWindowItemContext.

With this commit in place, the native GTK widgets
inside VCL toolbars now show up in Accerciser's
treeview of the LO a11y tree when using the gtk3
VCL plugin as expected.

Change-Id: I3b2182ca953056a939670e73687f4568620a4087
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182426
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>