tdf#165437: regression: CairoSDPR: Exclude SDPR when mirrored
This one is similar to tdf#165061 due to not using CairoSDPR when
RTL is used, but it seems that not all cases of 'mirroring' are
covered with using IsRTLEnabled. It also needs to check
HasMirroredGraphics, so I added that.

Change-Id: Ia86add7841e69657d8f64f7001eea4a57e927043
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182481
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>
(cherry picked from commit dedaef9565dba5856ef68a49c634b614cb3ef24d)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182504
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>