fix dodgy DrawController::fireChangeLayer code
wow. just wow.

So this "worked" previously by virture of the fact that we were storing
a pointer to something on the stack, and it would have fired almost all
of the time, so mostly the symptom would be unnecessary screen updates.

Of course, sometimes, by accident, two things on the stack would end up
in the same location and the firing would not happen, leading to a
temporary screen glitch.

Change-Id: Iea85520ce1a2f522a0d63a57fbd14017d1bb349c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182482
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>