fix dodgy lifecycle in OOoMountOperation
We fix/improve 3 things here
(1) Holding a pointer to something that might be on the stack seems dodgy.
I traced the calls back a few levels and ended up with lots of lots of
call sites, with no guarantee this pointer is safe to hold for any
length of time.

(2) we already have a unique_ptr in this structure, but we are not
properly constructing it.

(3) once we fix (1) and (2) by calling the constructor/destructor
explicitly, we can simply the password/username field handling.

This code has been here since:

    commit 065b374d695f5208c6dcdad8ead3f1ff6223036f
    Author: Vladimir Glazounov <vg@openoffice.org>
    Date:   Wed May 14 13:49:20 2008 +0000
    INTEGRATION: CWS ucpgio1 (1.1.2); FILE ADDED
    2008/04/30 16:35:19 cmc 1.1.2.1: #i88090# add gio content provider

Change-Id: I3ffcf30fbf66c855cedc4a3a99985bc6ee42faa9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182483
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins