iconchoicectrl a11y: Reuse parent handling + disposing from base class
Drop the custom handling for the accessible parent
from AccessibleIconChoiceCtrl, as the base class
VCLXAccessibleComponent already has an implementation
for vcl::Window subclasses that effectively does the same.

Drop AccessibleIconChoiceCtrl::disposing, which also
means that the base class one is used now, which
so far wasn't called from the override for no apparent
reason.

Change-Id: Iec4b297462fb8265e75fd9f298ba0838c23053db
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182491
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>