tdf#165059 sw fix not valid moveFrom/moveTo tag
According to the ISO standard, w:moveFrom/To elements are only permitted
within a move*RangeStart/End section. If not inside such delimited range
only w:ins and w:del are valid.

We still write w:move* in paragraph properties outside of move*Range.
This is technically also not valid, but lots of tests seem to have
documents with this, and it is outside the scope of the bug.

Change-Id: Iaf097b8b017aac1f6182a91abe464433e7664620
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182503
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>