cool#11254 desktop lok: avoid invalidations if no tiles are sent
If we have not rendered any tiles, then we should not need to emit
invalidations. To generalize that, for each view:
- when there is a clean-start or an EMPTY invalidation: reset a bbox of
  all tiles rendered
- whenever we render a new tile we should grow the bbox
- whenever we get an invalidate we should crop it against the bbox

This has to be per-view: view1 may have a different view render state
(e.g. spellcheck on) vs view2, so render in view1 should not influance
invalidations in view2. COOL has a concept of "canonical view ids" (a
group of views with the same render state), but that's not a problem for
us, as long as we always render on the same single view for the group,
the consistency is maintained.

For Calc we also have multiple parts, and Impress additionally have part
modes; the bbox should not be shared for these inside a single view.
Solve this by using a "part -> mode -> bbox" map of maps.

This then requires updating DesktopLOKTest::testBinaryCallback() in
CppunitTest_desktop_lib, because now some invalidates will be only
emitted if there was a paint before.

Change-Id: I5b18330ce56a0200a493d30ad51524bb46e1de02
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182524
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
(cherry picked from commit 43965ae570ded21653dabb5ec0570f8aaa7c497e)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182541
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>