tdf#161583 vcl,sc,sw: PDF/UA export: fix Contents of Link annotations
In ISO 14289-2:2024 the requirement was changed from "shall" to "should":

  Link annotations should include a Contents entry to enrich information
  available to assistive technology.

There is also now a Well-Tagged PDF (WTPDF) Version 1.0.0, which says:

  8.9.2.4.2 Link

  Link annotations should include a Contents entry to enrich information
  available to assistive technology.

  NOTE 1 Link annotations are often accessed out of context; the Contents
  entry provides optional additional information. The Contents entry is also
  particularly valuable in the context of link targets that are not intended
  to be human-readable.

  EXAMPLE — A link over the text “click here” is improved by a Contents entry
  to advise the user regarding the link’s target.

So adapt PDFWriterImpl::emitLinkAnnotations() to produce "Contents" only
if there is some alt text.

The sw a11y check will already warn if the alt text/"name" is missing on
a hyperlink in a SwTextNode, and a previous commit added a warning if it
is missing on a hyperlink in a SwFlyFrameFormat.

Contents should in any case not contain the textual representation of
the hyperlinks, that was a misunderstanding, but describe the target of
the link.

This requires adapting numerous unit tests.

(regression from commit fa3f04bdd4f73a1b3be70dfb709c44638ef7e3d9)

Change-Id: I88a7ae83d17d781115c93152d267ddb57208c200
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182600
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins
(cherry picked from commit 9cef7e5ba50b483cd84682528c7ef503965b6104)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182614
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>